apps/backend/src/modules/auth/auth.controller.ts
import { Controller, Get, Post, Body } from '@nestjs/common';
import { AuthService } from './auth.service';
import { CreateAuthDto } from './dto/create-auth.dto';
import { CurrentUser, Public } from '../../commons/guard';
import { ApiBearerAuth, ApiTags } from '@nestjs/swagger';

@Controller('auth')
@ApiTags('Auth')
export class AuthController {
  constructor(private readonly authService: AuthService) {}

  @Post()
  @Public()
  create(@Body() createAuthDto: CreateAuthDto) {
    return this.authService.create(createAuthDto);
  }

  @ApiBearerAuth()
  @Get('me')
  findOne(@CurrentUser() user: { id: number }) {
    return this.authService.findOne(user.id);
  }
}
